fix url encode in BaseRequest
